+++ conftest.py
@@ -56,22 +56,19 @@ def client(client, auth_backend, settings):
             }
         else:
             user_profile = {}
-        auth_backend.return_value = create_response(
-            200,
-            {
-                'id': user.id,
-                'email': user.email,
-                'hashed_uuid': user.hashed_uuid,
-                'user_profile': user_profile
-            }
-        )
+        auth_backend.return_value = create_response({
+            'id': user.id,
+            'email': user.email,
+            'hashed_uuid': user.hashed_uuid,
+            'user_profile': user_profile
+        })
     client.force_login = force_login
     return client
 
 
 @pytest.fixture(autouse=True)
 def mock_create_user_profile():
-    response = create_response(200, {
+    response = create_response({
         'first_name': 'First Name',
         'last_name': 'Last Name',
         'job_title': 'Director',