+++ core/tests/test_views.py
@@ -18,7 +18,7 @@ def test_companies_house_search_validation_error(client, settings):
 @mock.patch('core.views.ch_search_api_client.company.search_companies')
 def test_companies_house_search_api_error(mock_search, client, settings):
 
-    mock_search.return_value = create_response(400)
+    mock_search.return_value = create_response(status_code=400)
     url = reverse('api:companies-house-search')
 
     with pytest.raises(requests.HTTPError):
@@ -28,9 +28,7 @@ def test_companies_house_search_api_error(mock_search, client, settings):
 @mock.patch('core.views.ch_search_api_client.company.search_companies')
 def test_companies_house_search_api_success(mock_search, client, settings):
 
-    mock_search.return_value = create_response(
-        200, {'items': [{'name': 'Smashing corp'}]}
-    )
+    mock_search.return_value = create_response({'items': [{'name': 'Smashing corp'}]})
     url = reverse('api:companies-house-search')
 
     response = client.get(url, data={'term': 'thing'})
@@ -42,9 +40,7 @@ def test_companies_house_search_api_success(mock_search, client, settings):
 @mock.patch('core.views.ch_search_api_client.company.search_companies')
 def test_companies_house_search(mock_search, client, settings):
 
-    mock_search.return_value = create_response(
-        200, {'items': [{'name': 'Smashing corp'}]}
-    )
+    mock_search.return_value = create_response({'items': [{'name': 'Smashing corp'}]})
     url = reverse('api:companies-house-search')
 
     response = client.get(url, data={'term': 'thing'})
@@ -55,7 +51,7 @@ def test_companies_house_search(mock_search, client, settings):
 
 @mock.patch('core.views.requests.get')
 def test_address_lookup_bad_postcode(mock_get, client):
-    mock_get.return_value = create_response(400)
+    mock_get.return_value = create_response(status_code=400)
     url = reverse('api:postcode-search')
 
     response = client.get(url, data={'postcode': '21313'})
@@ -66,7 +62,7 @@ def test_address_lookup_bad_postcode(mock_get, client):
 
 @mock.patch('core.views.requests.get')
 def test_address_lookup_not_ok(mock_get, client):
-    mock_get.return_value = create_response(500)
+    mock_get.return_value = create_response(status_code=500)
     url = reverse('api:postcode-search')
 
     with pytest.raises(requests.HTTPError):
@@ -75,10 +71,7 @@ def test_address_lookup_not_ok(mock_get, client):
 
 @mock.patch('core.views.requests.get')
 def test_address_lookup_ok(mock_get, client):
-    mock_get    .return_value = create_response(200, {'addresses': [
-        '1 A road, , , , Ashire',
-        '2 B road, , , , Bshire',
-    ]})
+    mock_get.return_value = create_response({'addresses': ['1 A road, , , , Ashire', '2 B road, , , , Bshire']})
     url = reverse('api:postcode-search')
 
     response = client.get(url, data={'postcode': '123123'})