+++ profile/fab/helpers.py
@@ -1,6 +1,7 @@
 import http
 
 from directory_api_client.client import api_client
+from directory_constants import user_roles
 import directory_components.helpers
 
 
@@ -80,3 +81,14 @@ def disconnect_from_company(sso_session_id):
     response = api_client.supplier.disconnect_from_company(sso_session_id)
     response.raise_for_status()
     assert response.status_code == 200
+
+
+def create_admin_transfer_invite(sso_session_id, email):
+    response = api_client.company.create_transfer_invite(sso_session_id=sso_session_id, new_owner_email=email)
+    response.raise_for_status()
+    assert response.status_code == 201
+
+
+def is_sole_admin(sso_session_id):
+    collaborators = retrieve_collaborators(sso_session_id)
+    return [collaborator['role'] for collaborator in collaborators].count(user_roles.ADMIN) == 1