+++ profile/fab/tests/test_helpers.py
@@ -32,7 +32,7 @@ def test_profile_parser_no_data_serialize_for_template():
 @mock.patch.object(api_client.supplier, 'retrieve_profile')
 def test_get_supplier_profile(mock_retrieve_profile):
     data = {'name': 'Foo Bar'}
-    mock_retrieve_profile.return_value = create_response(200, data)
+    mock_retrieve_profile.return_value = create_response(data)
 
     profile = helpers.get_supplier_profile('1234')
 
@@ -43,7 +43,7 @@ def test_get_supplier_profile(mock_retrieve_profile):
 
 @mock.patch.object(api_client.supplier, 'retrieve_profile')
 def test_get_supplier_profile_not_found(mock_retrieve_profile):
-    mock_retrieve_profile.return_value = create_response(404)
+    mock_retrieve_profile.return_value = create_response(status_code=404)
 
     profile = helpers.get_supplier_profile('1234')
 
@@ -54,7 +54,7 @@ def test_get_supplier_profile_not_found(mock_retrieve_profile):
 
 @mock.patch.object(api_client.company, 'retrieve_private_profile')
 def test_get_company_profile_not_found(mock_retrieve_private_profile):
-    mock_retrieve_private_profile.return_value = create_response(404)
+    mock_retrieve_private_profile.return_value = create_response(status_code=404)
 
     profile = helpers.get_company_profile('1234')