+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - TT-1716 - Support removing collaborators
 - TT-1732 - Added context to incomplete business profile banner
 - TT-1713 - Allow user to remove themselves from the business profile
+- TT-1714 - Allow single collaborator to invite other admin
 
 ### Fixed bugs:
 - TT-1728 - Not ask personal details to individual upgrading to business profile